init.qcom.usb.sh: Add persist property check to enable multi UVC
Currently, boot KPIs are impacted due to the addition of multi UVC
support, which is not needed for all SPs.

Add a check based on the persist property
`persist.vendor.usb.multi.uvc`, which will be set on demand
when the use case is being tested and cleared by writing an empty
string (“”) to disable multi UVC support.

Change-Id: I10e2df87740e3d72fdbd554c697ae86daf5a3bb5